src/modules/users/domain/use-cases/register.ts
import { hash } from 'bcryptjs'
import { UserAlreadyExistsError } from '../errors/user-already-exist-error'
import { UserRepository } from '../repository/user-repository'
import { User } from '../models/user'

interface RegisterUserCaseRequest {
  name: string
  email: string
  password: string
}

interface RegisterUserUseCaseResponse {
  user: User
}

export class RegisterUserUseCase {
  constructor(private usersRepository: UserRepository) {}

  async execute({
    name,
    email,
    password,
  }: RegisterUserCaseRequest): Promise<RegisterUserUseCaseResponse> {
    const password_hash = await hash(password, 6)

    const userWithSameEmail = await this.usersRepository.findByEmail(email)

    if (userWithSameEmail) {
      throw new UserAlreadyExistsError()
    }

    const user = await this.usersRepository.create({
      name,
      email,
      password: password_hash,
    })
    return {
      user,
    }
  }
}
